add create class, create homework, join class, list member function

+++ student-manager/src/main/java/com/studentmanager/model/Submission.java
@@ -39,7 +39,7 @@ public class Submission {
     private String filePath;
 
     @Column
-    private Integer mark;
+    private Double mark;
 
     @Column
     @Builder.Default
